Cleaned up FileInfoNode
Separated node into node and UI node. The ui node being a partial class that complements the filenode. This for partial separation of tasks and a cleaner code overall I hope.
Also there was a redraw bug fixed for the progress bar since the readjustment ignored the adjusted height (we fixed that)